dtm0/recovery: Allow T-after-ONLINE in local rfom.
Problem: Hare may tell a Motr process that it enters TRANSIENT state
after the process has been recovered (entered ONLINE).  Formally
speaking, we have a case where a Motr process experiences a failure
during recovery (R -> T transition).  The reason is not clear but either
way DTM0 is not able to handle it at the moment. It assumes that Hare/HA
would kill the process if something goes wrong (instead of sending T).
Solution: Since we do not support consequent failures (R -> T),
we cannot guarantee correctness in this scenario. Therefore, we can just
release the restriction and allow the local recovery fom coroutine
to do nothing when it learns that its process enters TRANSIENT state.

Signed-off-by: Ivan Alekhin <ivan.alekhin@seagate.com>